Added Monte-Carlo Tree for Connect 4

+++ MiniMaxTree/MiniMaxTree/MiniMaxTree.cs
@@ -22,7 +22,7 @@ namespace MiniMaxTree
     {
         public MiniMaxNode<T>[] children;
         public T gameState;
-        public int Value;
+        public double Value;
 
         public MiniMaxNode(T gameState)
         {
